+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Atomic/AtomicGoThroughWaypoints_Snapshots/Test1.Whith_single_waypoint_waypoint=NodeSpy2.verified.mmd
