+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Atomic/AtomicGoThroughWaypoints_Snapshots/Test2.Whith_many_waypoints.verified.mmd
