+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Atomic/AtomicGotoDestination_Snapshots/Test2.When_origin_is_also_destination_with_waypoints.verified.mmd
