+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Chained/ChainedGoThroughWaypoints_Snapshots/Test1.Whith_single_waypoint_waypoint=Node3.verified.mmd
