+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Chained/ChainedGoThroughWaypoints_Snapshots/Test2.Whith_many_waypoints.verified.mmd
