+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Chained/ChainedGotoDestination_Snapshots/Test1.When_origin_is_also_destination_without_waypoints_or_explicit_self_reference_edge.verified.mmd
