+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Chained/ChainedUsesEdgesWithSmallestWeights_Snapshots/Test1.When_same_edge_with_different_weights.verified.mmd
