+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Chained/ChainedUsesEdgesWithSmallestWeights_Snapshots/Test2.When_edge_configured_with_default_values.verified.mmd
