+++ test/UnitTests/Caravel.Core.UnitTests/Tests/Navigation/Chained/ChainedUsesEdgesWithSmallestWeights_Snapshots/Test3.When_edges_configured_with_specific_and_default_values.verified.mmd
