+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Chained/ChainedChangesTheCurrentNode_Snapshots/Test1.When_action_returns_a_different_node.verified.mmd
