+++ test/UnitTests/Caravel.Core.UnitTests/Tests/NodeAction/Chained/ChainedHasDefaultJourneyAndLocalCancellationTokensToNone.cs
@@ -1,15 +1,15 @@
 ﻿using AwesomeAssertions;
 
-namespace Caravel.Core.UnitTests.Tests.NodeAction.Atomic;
+namespace Caravel.Core.UnitTests.Tests.NodeAction.Chained;
 
 [Trait(TestType, Unit)]
 [Trait(Feature, FeatureNodeAction)]
 [Trait(Domain, NodeDomain)]
-public sealed class HasDefaultJourneyAndLocalCancellationTokensToNone : IDisposable
+public sealed class ChainedHasDefaultJourneyAndLocalCancellationTokensToNone : IDisposable
 {
     private readonly CancellationTokenSource _localTokenSource30mins;
 
-    public HasDefaultJourneyAndLocalCancellationTokensToNone()
+    public ChainedHasDefaultJourneyAndLocalCancellationTokensToNone()
     {
         _localTokenSource30mins = new CancellationTokenSource(TimeSpan.FromMinutes(30));
     }
@@ -28,6 +28,9 @@ public sealed class HasDefaultJourneyAndLocalCancellationTokensToNone : IDisposa
             .WithEdge<Node2>()
             .Done()
             .AddNode<Node2>()
+            .WithEdge<Node3>()
+            .Done()
+            .AddNode<Node3>()
             .Done();
 
         // CancellationToken not set
@@ -35,8 +38,9 @@ public sealed class HasDefaultJourneyAndLocalCancellationTokensToNone : IDisposa
 
         // Act
         var sut = await journey
-            .DoAsync<Node1>((node, ct) => Task.FromResult(node))
-            .GotoAsync<Node2>(); // CancellationToken not set
+            .GotoAsync<Node2>()
+            .DoAsync<Node2>((node, ct) => Task.FromResult(node)) // CancellationToken not set
+            .GotoAsync<Node3>();
 
         // Assert
         journey.JourneyCancellationToken.IsCancellationRequested.Should().BeFalse();
